refactor: changing dir organization
