+++ src/components/NoProductFound/NoProductFound.tsx
@@ -1,7 +1,7 @@
 import { Image, View } from "react-native";
-import { Typography } from "../../Typography";
+import { Typography } from "../Typography";
 import { styles } from "./styles";
-import { Product } from "../../ProductSlide";
+import { Product } from "../ProductSlide";
 import interrogrationMark from "../../../../assets/images/interrogationMark.png";
 
 export const NoProductFound: React.FC = () => {
