+++ src/components/NoProductFound/styles.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import { colorPalette } from "../../../theme/colors";
+import { colorPalette } from "../../theme/colors";
 
 export const styles = StyleSheet.create({
   interrogationMark: {
