+++ src/components/Search/SearchBar/styles.ts
