+++ src/components/Search/SearchProductList/styles.ts
