+++ src/navigation/Navigation.tsx
@@ -2,7 +2,7 @@ import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import { HamburgerButton } from "../components/HamburgerButton";
-import { SearchBar } from "../components/SearchBar";
+import { SearchBar } from "../components/Search/SearchBar";
 import { SearchButton } from "../components/SearchButton";
 import { Dashboard } from "../screens/Dashboard";
 import { Home } from "../screens/Home";
